# Reducing memory requirements of 2N methods for special functions 