This is a very experimental PR. Just to demonstrate the trick to remove `k` from the 2N Low Memory Methods for some specific problem.
Trick is to make a wrapper of a register for which just overloads `du[i] = ...` to `du[i] += ...`. Using this the dependency of k is removed.

__Limitation__:
This won't work with the functions where du is on the right hand side. For example:
```julia
function f(du, u, p, t)
  du[1] = 5* u[1]
  du[2] = 2* du[1]
end
```
which should not be a problem as ,in theory, we write `du/dt = f(u, p, t)` so `du` can be written purely in terms of u, p and t. Still we can make a flag which will make sure from the user that this can be done.

I tried this code with my code and gave same output before and after the commit:
```julia
using OrdinaryDiffEq

function f(du,u,p,t)
	du[1] = 5 * u[2] + 4 * u[1]
	du[2] = 3 * u[2]
end
u0 = zeros(Float64, 2)
u0[1] = 1.0
u0[2] = 1.0

tspan = (0.0,1.0)
prob = ODEProblem(f,u0,tspan)
integ = init(prob, CarpenterKennedy2N54(),save_start=false,save_end=true,
                          save_everystep = false, calck = false, dt = 0.01)
solve!(integ)
@show(integ.sol)
```
If this idea is good to go,I would remove the unused registers and add the flag.