@ChrisRackauckas 2019-03-07
Yeah, I think a flag in each of the low memory algorithms about this optimization would be good to have. I would even have it set to true by default.