@kanav99 2019-03-08
Okay I will add more possibilities to the interface now. Been traveling, so would do it by evening (IST). This would need nice documentation too.