@kanav99 2019-03-09
Current status:
Last commit successfully removes one register from the cache. Current naming conventions are a bit misleading (tmp is actually similar to fsal) and the flag is not yet done. Output is correct too for basic problems. Needs a lot of work and testing to verify that it covers all the cases.