@kanav99 2019-03-13
Actual work is almost done. The only problem is that while using `.` I am getting large number of extra allocations. I think that this is problem is solvable, just the midway statements aren't getting optimized somewhere, but I currently don't know how to proceed, tried many ways, but this was the best method I came upon. I would need help of someone who have implemented custom and in-place broadcast before. Requesting reviews @ChrisRackauckas @YingboMa 

*Benchmarks*:
Script:
```
function f(du,u,p,t)
 du .= u
end

u0 = [1.0, 1.0]
tspan = (0.0,1.0)
prob = ODEProblem(f,u0,tspan)
integ = init(prob, CarpenterKennedy2N54(),save_start=false,save_end=false,
                          save_everystep = false, calck = false, dt = 0.01, alias_u0 = true)
@time solve!(integ)
integ = init(prob, CarpenterKennedy2N54(),save_start=false,save_end=true,
                          save_everystep = false, calck = true, dt = 0.01)
@time solve!(integ)
integ = init(prob, CarpenterKennedy2N54(),save_start=false,save_end=true,
                          save_everystep = false, calck = true, dt = 0.01)
@time solve!(integ)
@show(integ.sol)
@show(u0)


u0 = rand(1000,1000)
prob = ODEProblem(f,u0,tspan)
integ = init(prob, CarpenterKennedy2N54(),save_start=false,save_end=false,
                          save_everystep = false, calck = false, dt = 0.1, alias_u0 = true)

@show Base.summarysize(integ)/Base.summarysize(u0)
```

Master:
```
  0.979325 seconds (1.99 M allocations: 97.611 MiB, 4.47% gc time)
  0.000027 seconds (6 allocations: 336 bytes)
  0.000027 seconds (6 allocations: 336 bytes)
integ.sol = retcode: Success
Interpolation: 1st order linear
t: [1.0]
u: Array{Float64,1}[[7.38906, 7.38906]]
u0 = [2.71828, 2.71828]
Base.summarysize(integ) / Base.summarysize(u0) = 3.000164374178129
```
With flag turned to true:
```
  0.936926 seconds (1.75 M allocations: 85.362 MiB, 3.95% gc time)
  0.000043 seconds (406 allocations: 12.844 KiB)
  0.000026 seconds (406 allocations: 12.844 KiB)
integ.sol = retcode: Success
Interpolation: 1st order linear
t: [1.0]
u: Array{Float64,1}[[7.38906, 7.38906]]
u0 = [2.71828, 2.71828]
Base.summarysize(integ) / Base.summarysize(u0) = 2.0001653741731293
```
Flag = false:
```
  1.466949 seconds (1.84 M allocations: 89.197 MiB, 3.92% gc time)
  0.000030 seconds (9 allocations: 608 bytes)
  0.000029 seconds (9 allocations: 608 bytes)
integ.sol = retcode: Success
Interpolation: 1st order linear
t: [1.0]
u: Array{Float64,1}[[7.38906, 7.38906]]
u0 = [2.71828, 2.71828]
Base.summarysize(integ) / Base.summarysize(u0) = 2.0001653741731293
```