@ChrisRackauckas 2019-03-13
The tests need to be updated to make sure it's only 2 registers.