@kanav99 2019-03-15
Added some more of the tests to test both with and without the flag. Have started documentation for this.