@ChrisRackauckas 2019-03-15
lgtm. Still WIP?